cleaned up, eliminated unnecessary combined reducer setup
