--- client/store/project.js
@@ -1,31 +1,0 @@
-// // export const FILTER_PROJECTS = 'FILTER_PROJECTS'
-
-// // export const filterProjects = (projArr, catObj) => ({
-// //   type: FILTER_PROJECTS,
-// //   projArr, catObj
-// // })
-
-// // export const filteringProjects = (projArr, catObj) => {
-// //   return dispatch => {
-// //     let currentProjects = projArr.filter(p => catObj[p.categoryId])
-// //     console.log('************* currentProjects', currentProjects)
-// //     dispatch(filterProjects(currentProjects))
-// //   }
-// // }
-
-// const initialState = {
-//   projects: []
-// }
-
-// const project = (state = initialState, action) => {
-//   switch (action.type) {
-//     // case FILTER_PROJECTS:
-//     //   console.log('****************', action)
-//     // return {...state, projects: action.projArr.filter(p => action.catObj[p.categoryId])}
-//     //   // return {...state, projects: action.projects}
-//     default:
-//       return state
-//   }
-// }
-
-// export default project
